+++ electricitymap/contrib/parsers/ENTSOE.py
@@ -272,7 +272,6 @@ ENTSOE_PRICE_DOMAIN_MAPPINGS: dict[str, str] = {
     "DE": ENTSOE_DOMAIN_MAPPINGS["DE-LU"],
     "IE": ENTSOE_DOMAIN_MAPPINGS["IE(SEM)"],
     "GB-NIR": ENTSOE_DOMAIN_MAPPINGS["IE(SEM)"],
-    "LU": ENTSOE_DOMAIN_MAPPINGS["DE-LU"],
     "UA": ENTSOE_DOMAIN_MAPPINGS["UA-IPS"],
 }
