+++ electricitymap/contrib/parsers/tests/test_ENTSOE_price_overrides.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import pytest
 from requests_mock import ANY, GET
 
-from electricitymap.contrib.parsers.AX import fetch_price
+from electricitymap.contrib.parsers.ENTSOE_price_overrides import fetch_price
 from electricitymap.contrib.types import ZoneKey
 
 base_path_to_mock = Path("electricitymap/contrib/parsers/tests/mocks/ENTSOE")
@@ -15,11 +15,12 @@ def entsoe_token_env():
     os.environ["ENTSOE_TOKEN"] = "token"
 
 
-def test_fetch_price(requests_mock, session, snapshot):
+@pytest.mark.parametrize("zone", ["AX", "LU"])
+def test_fetch_price(requests_mock, session, snapshot, zone):
     data = base_path_to_mock / "FR_prices.xml"
     requests_mock.register_uri(
         GET,
         ANY,
         content=data.read_bytes(),
     )
-    assert snapshot == fetch_price(ZoneKey("AX"), session)
+    assert snapshot == fetch_price(ZoneKey(zone), session)
